Trait ironsea_index::RecordFields

pub trait RecordFields<F> {
    fn fields(&self) -> F;
}

Record behavior used by IndexedDestructured implementations.

RecordFields is used by indices which de-structure records into two components, the key and the fields associated to that unique key. This is provided by the users of indices, for each of their struct they wish to index.

Required methods

fn fields(&self) -> F

Extract the fields of the record

Loading content...